+++ R/markets.R
@@ -2,12 +2,14 @@
 #'
 #' Get data.frame with prices of cryptocurrencies using the REST API of cryptowat.ch.
 #'
-#' @usage markets(pair, params = NULL, exchange = "kraken", datetime = TRUE, api_key = NULL)
-#' @param pair A string containing a pair symbol, e.g. \emph{btcusd} (required argument).
+#' @usage markets(pair, params = NULL, exchange = "kraken", datetime = TRUE,
+#'                api_key = NULL, allowance = FALSE)
+#' @param pair A string containing a pair symbol, e.g. \emph{btcusd} (required argument). Run \code{get_pairs()} to find other available pairs.
 #' @param params A list containing \code{before}, \code{after} and \code{periods} which is measured in seconds (optional). See \url{https://docs.cryptowat.ch/rest-api/markets/ohlc} for further information.
-#' @param exchange A string containing the exchange. Default is \emph{kraken}. Run \code{get_exchanges()} to choose other exchanges.
+#' @param exchange A string containing the exchange. Default is \emph{kraken}. Run \code{get_exchanges()} to find other available exchanges.
 #' @param datetime A logical. \code{TRUE} indicates that datetime type is used. \code{FALSE} indicates \emph{unix timestamp}. Default is \code{TRUE}.
 #' @param api_key A string containing the API key. See \url{https://docs.cryptowat.ch/rest-api/rate-limit} to learn how to create an account and how to generate an API key.
+#' @param allowance Logical (default is \emph{FALSE}). If \emph{TRUE} the function returns a list which includes allowance information, i.e. cost of the request, remaining credits and your account name.
 #'
 #' @return A data.frame containing prices of a given pair of currencies.
 #'
@@ -16,7 +18,7 @@
 #' @examples
 #' \dontrun{
 #' # Prices (every four hours) of Bitcoin in USD
-#' df.btcusd. <- markets("btcusd")
+#' df.btcusd <- markets("btcusd")
 #' # Hourly prices of Bitcoin in USD for a specific period
 #' df.btcusd.hourly <- markets("btcusd", list(periods = 3600, before = 1609851600,
 #'                                           after = 1609506000), datetime = FALSE)
@@ -26,25 +28,37 @@
 #' }
 #'
 #' @export
-markets <- function(pair, params = NULL, exchange = "kraken", datetime = TRUE, api_key = NULL) {
+markets <- function(pair, params = NULL, exchange = "kraken", datetime = TRUE, api_key = NULL, allowance = FALSE) {
 
   route <- "ohlc"
 
   if (datetime) {
 
-    if (!is.null(params[["before"]])) params$before <- as.numeric(as.POSIXct(params$before))
-    if (!is.null(params[["after"]])) params$after <- as.numeric(as.POSIXct(params$after))
+    if ( !is.null(params[["before"]]) ) params$before <- as.numeric(as.POSIXct(params$before))
+    if ( !is.null(params[["after"]]) ) params$after <- as.numeric(as.POSIXct(params$after))
 
   }
 
-  prices <- get_markets(route, pair, exchange, params, api_key)
+  prices <- get_markets(route, pair, exchange, params, api_key, allowance)
 
-  df.prices <- data.frame(prices[[1]][[1]])
+  if (allowance) {
+
+    df.prices <- data.frame(prices[[1]][[1]])
+
+    allowance.data <- prices[[2]]
+
+  } else if (allowance == FALSE) {
+
+    df.prices <- data.frame(prices[[1]])
+
+  }
 
   names(df.prices) <- c("CloseTime", "OpenPrice", "HighPrice", "LowPrice", "ClosePrice", "Volume", "QuoteVolume")
 
   if (datetime) df.prices$CloseTime <- lubridate::as_datetime(df.prices$CloseTime)
 
-  return(df.prices)
+  if (allowance) { output <- list(result = df.prices, allowance = allowance.data) } else if (allowance == FALSE) { output <- df.prices }
+
+  return(output)
 
 }