+++ NEWS.md
@@ -1,7 +1,8 @@
 # cryptowatchR 0.1.0.9000
 
-* Added the possibility to use your own API key for all functions. See argument `api_key`.
 * Now `get_markets()` works with `summaries` and `prices`
+* Added the possibility to use your own API key for all functions. See argument `api_key`.
+* Added the possibility to show allowance (costs per request, remaining credits). See argument `allowance`.
 
 # cryptowatchR 0.1.0